Add files via upload
create assignment module;
Includes variables from an html form that captures assignmentId, start time and date as well as the submission time and date. 
From this, the total duration of availability is calculated using the js getTime() function.
the form also captures the characters to be included in the assessment.